add sample code of navigation
