+++ screens/DetailScreen.tsx
@@ -1,0 +1,9 @@
+import { Text, View } from 'react-native';
+
+export default function DetailScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Details Screen</Text>
+    </View>
+  );
+}
