+++ types/Navigation.ts
@@ -1,0 +1,8 @@
+import type { NativeStackScreenProps } from '@react-navigation/native-stack';
+
+export type RootStackParamList = {
+  Home: undefined;
+  Detail: undefined;
+};
+
+export type NavigationProps = NativeStackScreenProps<RootStackParamList, 'Home', 'Detail'>;
